+++ recipe/oauth2provider/index.d.ts
@@ -1,4 +1,4 @@
-export * from "../../../lib/build/recipe/oauth2/types";
+export * from "../../lib/build/recipe/oauth2provider";
 /**
  * 'export *' does not re-export a default.
  * import NextJS from "supertokens-node/nextjs";
@@ -6,5 +6,5 @@ export * from "../../../lib/build/recipe/oauth2/types";
  * - user add "esModuleInterop": true in their tsconfig.json file
  * - we do the following change:
  */
-import * as _default from "../../../lib/build/recipe/oauth2/types";
+import * as _default from "../../lib/build/recipe/oauth2provider";
 export default _default;