+++ test/test-server/src/index.ts
@@ -19,8 +19,8 @@ import ThirdPartyRecipe from "../../../lib/build/recipe/thirdparty/recipe";
 import { TypeInput as ThirdPartyTypeInput } from "../../../lib/build/recipe/thirdparty/types";
 import { TypeInput as MFATypeInput } from "../../../lib/build/recipe/multifactorauth/types";
 import TOTPRecipe from "../../../lib/build/recipe/totp/recipe";
-import OAuth2Recipe from "../../../lib/build/recipe/oauth2/recipe";
-import { TypeInput as OAuth2TypeInput } from "../../../lib/build/recipe/oauth2/types";
+import OAuth2ProviderRecipe from "../../../lib/build/recipe/oauth2provider/recipe";
+import { TypeInput as OAuth2ProviderTypeInput } from "../../../lib/build/recipe/oauth2provider/types";
 import OAuth2ClientRecipe from "../../../lib/build/recipe/oauth2client/recipe";
 import { TypeInput as OAuth2ClientTypeInput } from "../../../lib/build/recipe/oauth2client/types";
 import { TypeInput as OpenIdRecipeTypeInput } from "../../../lib/build/recipe/openid/types";
@@ -37,7 +37,7 @@ import Session from "../../../recipe/session";
 import { verifySession } from "../../../recipe/session/framework/express";
 import ThirdParty from "../../../recipe/thirdparty";
 import TOTP from "../../../recipe/totp";
-import OAuth2 from "../../../recipe/oauth2";
+import OAuth2Provider from "../../../recipe/oauth2provider";
 import OAuth2Client from "../../../recipe/oauth2client";
 import accountlinkingRoutes from "./accountlinking";
 import emailpasswordRoutes from "./emailpassword";
@@ -46,7 +46,7 @@ import { logger } from "./logger";
 import multiFactorAuthRoutes from "./multifactorauth";
 import multitenancyRoutes from "./multitenancy";
 import passwordlessRoutes from "./passwordless";
-import oAuth2Routes from "./oauth2";
+import OAuth2ProviderRoutes from "./oauth2provider";
 import sessionRoutes from "./session";
 import supertokensRoutes from "./supertokens";
 import thirdPartyRoutes from "./thirdparty";
@@ -92,7 +92,7 @@ function STReset() {
     ProcessState.getInstance().reset();
     MultiFactorAuthRecipe.reset();
     TOTPRecipe.reset();
-    OAuth2Recipe.reset();
+    OAuth2ProviderRecipe.reset();
     OAuth2ClientRecipe.reset();
     SuperTokensRecipe.reset();
 }
@@ -298,8 +298,8 @@ function initST(config: any) {
                 })
             );
         }
-        if (recipe.recipeId === "oauth2") {
-            let initConfig: OAuth2TypeInput = {
+        if (recipe.recipeId === "oauth2provider") {
+            let initConfig: OAuth2ProviderTypeInput = {
                 ...config,
             };
             if (initConfig.override?.functions) {
@@ -314,7 +314,7 @@ function initST(config: any) {
                     apis: getFunc(`${initConfig.override.apis}`),
                 };
             }
-            recipeList.push(OAuth2.init(initConfig));
+            recipeList.push(OAuth2Provider.init(initConfig));
         }
         if (recipe.recipeId === "oauth2client") {
             let initConfig: OAuth2ClientTypeInput = {
@@ -425,7 +425,7 @@ app.use("/test/multifactorauth", multiFactorAuthRoutes);
 app.use("/test/thirdparty", thirdPartyRoutes);
 app.use("/test/totp", TOTPRoutes);
 app.use("/test/usermetadata", userMetadataRoutes);
-app.use("/test/oauth2", oAuth2Routes);
+app.use("/test/oauth2provider", OAuth2ProviderRoutes);
 
 // *** Custom routes to help with session tests ***
 app.post("/create", async (req, res, next) => {